+++ src/TitleBindPieces/TitleBindPiecesContainer/TitleBindPiecesContainer.js
@@ -89,8 +89,14 @@ export const TitleBindPiecesContainer = () => {
   const bindItems = (requestData) => {
     return bindPieces(requestData)
       .then(() => {
+        const barcode = requestData.bindItem?.barcode;
+        const messageId = barcode ? 'withBarcode' : 'noBarcode';
+
         onCancel();
-        showCallout({ messageId: 'ui-receiving.bind.pieces.create.success' });
+        showCallout({
+          messageId: `ui-receiving.bind.pieces.create.success.${messageId}`,
+          values: { barcode },
+        });
       }).catch((error) => handleMutationError(error));
   };