+++ src/common/hooks/useTitle/useTitle.js
@@ -13,7 +13,7 @@ export const useTitle = (titleId, options = {}) => {
     tenantId,
   } = options;
 
-  const ky = useOkapiKy({ tenantId });
+  const ky = useOkapiKy({ tenant: tenantId });
   const [namespace] = useNamespace('receiving-title');
 
   const {